requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
dnspython==2.2.1
fastapi==0.62.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.14.10
psycopg2-binary==2.9.3
pydantic==1.9.0
PyJWT==1.7.1
python-decouple==3.3
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
starlette==0.13.6
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.12.3
